email_mirror: Rename the email_mirror.py files to email_gateway.py.
This is step 1 of a naming migration we wanted to do. This commit just
handles renaming the python files and there is various "email mirror"
naming left over, so this requires follow up commits.
