+++ puppet/zulip/files/cron.d/email-mirror
@@ -1,3 +1,3 @@
 MAILTO=root
 
-* * * * *   zulip cd /home/zulip/deployments/current && ./manage.py email_mirror
+* * * * *   zulip cd /home/zulip/deployments/current && ./manage.py email_gateway
