+++ tools/linter_lib/custom_check.py
@@ -41,10 +41,10 @@ FILES_WITH_LEGACY_SUBJECT = {
 
     # These use subject in the email sense, and will
     # probably always be exempt:
-    'zerver/lib/email_mirror.py',
+    'zerver/lib/email_gateway.py',
     'zerver/lib/feedback.py',
     'zerver/tests/test_new_users.py',
-    'zerver/tests/test_email_mirror.py',
+    'zerver/tests/test_email_gateway.py',
 
     # These are tied more to our API than our DB model.
     'zerver/openapi/python_examples.py',
